+++ header/builtin_func.h
@@ -11,13 +11,37 @@ typedef enum e_move_position
 	MOVE_TO_OLDPWD
 }	t_move_position;
 
-void builtin_echo(t_condition *condition, char **argv);
-void builtin_env(t_condition *condition);
-void builtin_export(t_condition *condition, char **argv);
-void builtin_unset(t_condition *condition, char **argv);
-int builtin_pwd(t_condition *cond, char **argv);
-int builtin_cd(t_condition *cond, char **args);
-int	builtin_exit(t_condition *condition, char **args);
-char *get_item_value(t_item *item, char *key);
+void	builtin_echo(t_condition *condition, char **argv);
+void	builtin_env(t_condition *condition);
+void	builtin_export(t_condition *condition, char **argv);
+void	builtin_unset(t_condition *condition, char **argv);
+int		builtin_pwd(t_condition *cond, char **argv);
+int		builtin_cd(t_condition *cond, char **args);
+int		builtin_exit(t_condition *condition, char **args);
+char	*get_item_value(t_item *item, char *key);
+
+//cd_until
+int				update_cwd(t_condition *cond, char *newcwd);
+char			*lst_getenv(t_item *item, char *key);
+void			put_cd_error(t_condition *cond, char *str, char *perr);
+int		move_path(int option, t_condition cond);
+int		update_old_pwd(t_condition *cond);
+char			*get_item_value(t_item *item, char *key);
+int				update_item_value(t_condition *cond, t_item *item, const char *cwd);
+int				builtin_cd(t_condition *cond, char **args);
+//exit_until
+void	numeric_argument_error(char *argment);
+int		tma_error_check(t_condition *condition, char **args);
+bool	is_spase(char c);
+char	*skip_space(char *str);
+int		get_sign_skip0(char **str);
+//exit_until2
+long	over_256(long num);
+long	negative_num(long num);
+long	exit_status_num(long num);
+bool	is_check_num(char *str);
+//export_until.c
+bool	is_envname(char c);
+void	put_export_error(t_condition *condition, char *argv);
 
 # endif
