+++ src/builtin_func/builtin_echo.c
@@ -2,24 +2,32 @@
 #include "../../header/condition.h"
 #include "../../libft/libft.h"
 
-void 	builtin_echo(t_condition *condition, char **argv)
+bool	echo_n_option(char *argment)
+{
+	int	j;
+
+	j = 2;
+	if (argment && !ft_strncmp(argment, "-n", 2))
+	{
+		while (argment[j] == 'n')
+			j++;
+		if (argment[j] == '\0')
+			return (true);
+	}
+	return (false);
+}
+
+void	builtin_echo(t_condition *condition, char **argv)
 {
 	int		i;
-	int		j;
 	bool	is_newline;
 
 	i = 1;
 	is_newline = true;
-	if (argv[1] && !ft_strncmp(argv[1], "-n", 2))
+	if (echo_n_option(argv[1]))
 	{
-		j = 2;
-		while (argv[1][j] == 'n')
-			j++;
-		if (argv[1][j] == '\0')
-		{
-			is_newline = false;
-			i++;
-		}
+		is_newline = false;
+		i++;
 	}
 	while (argv[i])
 	{
