+++ src/builtin_func/builtin_env.c
@@ -16,5 +16,4 @@ void	builtin_env(t_condition *condition)
 		ft_putstr_fd("\n", STDOUT_FILENO);
 		current = current->next;
 	}
-
 }
